Parsing LBA1 grids
